Move releases table components to separate folder
